some last change

+++ CSRF_Node/src/users/controller/user.js
@@ -12,13 +12,28 @@ async function getUsers(req,res){
 }
 
 async function deleteUser(req, res) {
+
     try {
+
+    const FromHeader = req.headers["X-CSRF-Token"];
+    console.log("Delete request for name in header:", nameFromHeader);
+
+    
+    // const nameFromCookies = req.cookies["csrf-token"];
+    // console.log(nameFromCookies,"cookies========")
+
+    
+    if (FromHeader !== FromCookies) {
+      return res.status(403).json({ message: "Unauthorized: User mismatch" });
+    }
+
       const { name } = req.params; // Get the name from URL params
       console.log("Delete request for name:", name);
       
       const deletedUser = await userService.deleteUserByName(name);
       
       res.json({ message: `User ${name} deleted successfully`, user: deletedUser });
+
     } catch (error) {
       console.error(error);
       res.status(500).json({ message: error.message });
@@ -27,8 +42,3 @@ async function deleteUser(req, res) {
 
 
 module.exports={ getUsers,deleteUser }
-
-/**
-  Service me bussiness logic (kya kaam kar wana hai )
-  Controller 
- */
